feat: add fully developed nav/menu
Add a menu with all necessary links, links must differ for
the signed/unsigned user. Add the MUI nav button inside of the
log's route template, before the children. The Nav button is an MUI Button
with a Link component prop that leads to the home page, except one time: "log/create" route.